rename from handler to controller and cleanup some more
